requirements.txt
flask>=2.0.0
flask-socketio>=5.1.1
python-engineio>=4.0.0
python-socketio>=5.0.0
opencv-python-headless>=4.5.0
numpy>=1.19.0
Pillow>=9.2.0
pyautogui>=0.9.53
python-xlib>=0.31
pynput>=1.7.6
sounddevice>=0.4.4
websockets>=10.0
aiohttp>=3.8.0
python-engineio>=4.3.4
eventlet>=0.33.0
PyAudio>=0.2.13 